Support nanosecond timestamps in parquet (#10063)
Closes https://github.com/rapidsai/cudf/issues/9393

This PR was intended to support nanoseconds for both duration and timestamp types in parquet. It introduces `LogicalType`-handling on both reader and writer sides. This PR also includes code cleanups like moving `CompactProtocolReader` to its own file. Finally, nanosecond durations remain unchanged since it's not fully supported by `pyarrow` i.e. nanosecond durations are truncated to microseconds (see [here](https://github.com/apache/arrow/blob/release-7.0.0/cpp/src/arrow/python/datetime.cc#L259-L264)).

Authors:
  - Yunsong Wang (https://github.com/PointKernel)

Approvers:
  - Robert Maynard (https://github.com/robertmaynard)
  - Vukasin Milovanovic (https://github.com/vuule)
  - Devavret Makkar (https://github.com/devavret)
  - Karthikeyan (https://github.com/karthikeyann)

URL: https://github.com/rapidsai/cudf/pull/10063
