+++ cpp/src/io/parquet/reader_impl.cu
@@ -21,6 +21,8 @@
 
 #include "reader_impl.hpp"
 
+#include "compact_protocol_reader.hpp"
+
 #include <io/comp/gpuinflate.h>
 #include <io/utilities/config_utils.hpp>
 #include <io/utilities/time_utils.cuh>
@@ -105,13 +107,14 @@ type_id to_type_id(SchemaElement const& schema,
                    bool strings_to_categorical,
                    type_id timestamp_type_id)
 {
-  parquet::Type physical                = schema.type;
-  parquet::ConvertedType converted_type = schema.converted_type;
-  int32_t decimal_scale                 = schema.decimal_scale;
+  parquet::Type const physical            = schema.type;
+  parquet::LogicalType const logical_type = schema.logical_type;
+  parquet::ConvertedType converted_type   = schema.converted_type;
+  int32_t decimal_scale                   = schema.decimal_scale;
 
   // Logical type used for actual data interpretation; the legacy converted type
   // is superceded by 'logical' type whenever available.
-  auto inferred_converted_type = logical_type_to_converted_type(schema.logical_type);
+  auto const inferred_converted_type = logical_type_to_converted_type(logical_type);
   if (inferred_converted_type != parquet::UNKNOWN) converted_type = inferred_converted_type;
   if (inferred_converted_type == parquet::DECIMAL && decimal_scale == 0)
     decimal_scale = schema.logical_type.DECIMAL.scale;
@@ -130,12 +133,12 @@ type_id to_type_id(SchemaElement const& schema,
     case parquet::TIME_MICROS:
       return (timestamp_type_id != type_id::EMPTY) ? timestamp_type_id
                                                    : type_id::DURATION_MICROSECONDS;
-    case parquet::TIMESTAMP_MICROS:
-      return (timestamp_type_id != type_id::EMPTY) ? timestamp_type_id
-                                                   : type_id::TIMESTAMP_MICROSECONDS;
     case parquet::TIMESTAMP_MILLIS:
       return (timestamp_type_id != type_id::EMPTY) ? timestamp_type_id
                                                    : type_id::TIMESTAMP_MILLISECONDS;
+    case parquet::TIMESTAMP_MICROS:
+      return (timestamp_type_id != type_id::EMPTY) ? timestamp_type_id
+                                                   : type_id::TIMESTAMP_MICROSECONDS;
     case parquet::DECIMAL:
       if (physical == parquet::INT32) { return type_id::DECIMAL32; }
       if (physical == parquet::INT64) { return type_id::DECIMAL64; }
@@ -159,6 +162,12 @@ type_id to_type_id(SchemaElement const& schema,
     case parquet::NA: return type_id::STRING;
     // return type_id::EMPTY; //TODO(kn): enable after Null/Empty column support
     default: break;
+  }
+
+  if (inferred_converted_type == parquet::UNKNOWN and physical == parquet::INT64 and
+      logical_type.TIMESTAMP.unit.isset.NANOS) {
+    return (timestamp_type_id != type_id::EMPTY) ? timestamp_type_id
+                                                 : type_id::TIMESTAMP_NANOSECONDS;
   }
 
   // is it simply a struct?
@@ -1699,6 +1708,7 @@ table_with_metadata reader::impl::read(size_type skip_rows,
                                            required_bits(schema.max_repetition_level),
                                            col_meta.codec,
                                            converted_type,
+                                           schema.logical_type,
                                            schema.decimal_scale,
                                            clock_rate,
                                            i,
